respect checksum header over config setting
